refactor(utils): #550 — split utils.py into 5 domain modules; delete utils.py
Closes #550. Final child of M4 (#545); closes the M4 epic.

Mechanical split — no logic changes inside any moved function. utils.py's
593 LOC are partitioned across 5 new domain modules + 2 outliers, then
utils.py itself is deleted (no backwards-compat shim per CLAUDE.md
no-backwards-compat-hacks rule).

New modules in src/findajob/:
- audit.py       — log_event, write_audit, LOG_PATH
- profile.py     — read_candidate_name, read_file_prefix,
                   _clean_profile_field, load_voice_samples
- prep_naming.py — safe_filename_part, build_prep_filenames,
                   build_outreach_filename
- classification.py — is_aggregator_company, is_valid_company,
                      is_ingest_noise_title, is_synthetic_job,
                      jd_is_usable, strip_jd_boilerplate, JD_MAX_CHARS
- llm_parsing.py — extract_json_payload, validate_llm_json

Outliers placed in existing modules:
- findajob.prep.quarantine.quarantine_stale_prep_folders — folder
  maintenance is prep-domain, not a cross-cutting utility.
- findajob.paths.load_env — env-file reading lives next to BASE
  resolution and config/paths.env loading. Lets utils.py be deleted
  entirely instead of left as a 1-function stub.

Cross-module dependency: log_event (now in findajob.audit) is called
by quarantine_stale_prep_folders and strip_jd_boilerplate. Both new
modules `from findajob.audit import log_event`.

51 files updated:
- 38 inbound imports rewritten to point at new modules.
- 5 new module files created.
- 1 file deleted (src/findajob/utils.py).
- 7 test files updated their `from findajob import utils` patterns to
  `from findajob import audit` (covering monkeypatch.setattr targets
  for LOG_PATH and load_env in import-safety tests).

ruff --fix consolidated multi-symbol imports across the 4 orchestrator
modules (interview, prep, triage, find_contacts) and merged
`from findajob.paths import BASE` lines with the new
`from findajob.paths import load_env` additions.

Verification:
- src/findajob/utils.py does not exist.
- git grep -nE 'findajob\\.utils|from findajob import .*\\butils\\b'
  on src/, scripts/, tests/ returns empty.
- mypy + ruff check + ruff format --check + full pytest (1927 passed,
  1 skipped pre-existing) all green locally.

Net diff: +128 / -710 across 51 files. The codebase is smaller after
this refactor: 5 focused modules totaling 593 LOC replace one 593-LOC
junk drawer, but ruff's import-organize pass condensed multi-line
imports across many call sites.
