+++ scripts/ingest_form.py
@@ -31,10 +31,10 @@ from datetime import UTC, datetime
 from google.oauth2 import service_account
 from googleapiclient.discovery import build
 
+from findajob.audit import log_event
 from findajob.cleaning import fingerprint, is_coarse_location, loose_fingerprint
 from findajob.db import connect
-from findajob.paths import BASE
-from findajob.utils import load_env, log_event
+from findajob.paths import BASE, load_env
 
 DB_PATH = f"{BASE}/data/pipeline.db"
 SA_FILE = f"{BASE}/config/gsheets_creds.json"
