+++ scripts/watchdog.py
@@ -25,9 +25,9 @@ from datetime import UTC, datetime, timedelta
 from pathlib import Path
 
 from findajob.actions import reset_prep_to_scored
+from findajob.audit import log_event
 from findajob.db import connect
 from findajob.paths import BASE
-from findajob.utils import log_event
 
 DB_PATH = f"{BASE}/data/pipeline.db"
 STALE_PREP_MINUTES = 60
