+++ src/findajob/actions.py
@@ -19,8 +19,8 @@ import sys
 from datetime import UTC, datetime
 from typing import Any
 
+from findajob.audit import log_event, write_audit
 from findajob.paths import BASE
-from findajob.utils import log_event, write_audit
 
 
 def handle_rejection(conn: sqlite3.Connection, job: Any, reason: str) -> bool:
@@ -35,7 +35,7 @@ def handle_rejection(conn: sqlite3.Connection, job: Any, reason: str) -> bool:
     # jd_excerpt: first 500 chars of raw_jd_text for post-hoc analysis
     jd = conn.execute("SELECT raw_jd_text, prep_folder_path FROM jobs WHERE id=?", (job["id"],)).fetchone()
     jd_excerpt = (jd["raw_jd_text"] or "")[:500] if jd and jd["raw_jd_text"] else ""
-    from findajob.utils import is_synthetic_job  # local import to avoid circular at module load
+    from findajob.classification import is_synthetic_job  # local import to avoid circular at module load
 
     if not is_synthetic_job(job):
         conn.execute(
