+++ src/findajob/fetchers/adapters/jobs_api14.py
@@ -8,8 +8,8 @@ from typing import ClassVar
 
 import requests
 
+from findajob.audit import log_event
 from findajob.cleaning import clean_company, clean_title
-from findajob.utils import log_event
 
 from ._keys import resolve_rapidapi_key
 from ._locations import read_target_locations
